training endpoint + javascript code for the training page
